[64 BITS] * NEW: Continue on fixing the use of long datatype.
git-svn-id: svn://localhost/gambas/trunk@917 867c0c6c-44f3-4631-809d-bfa615b0a4ec

+++ main/gbc/gb_error.h
@@ -39,7 +39,7 @@
 
 typedef
   struct {
-    long code;
+    int code;
     char msg[MAX_ERROR_MSG + 1];
     }
   ERROR_INFO;
